+++ params/convert/pyethereum.go
@@ -13,7 +13,7 @@ import (
 // chain specification format.
 func NewPyEthereumGenesisSpec(network string, genesis *paramtypes.Genesis) (*pyethereum.PyEthereumGenesisSpec, error) {
 	// Only ethash is currently supported between go-ethereum and pyethereum
-	if genesis.Config.Ethash == nil {
+	if genesis.Config.(*paramtypes.ChainConfig).Ethash == nil {
 		return nil, errors.New("unsupported consensus engine")
 	}
 	spec := &pyethereum.PyEthereumGenesisSpec{
